requirements.txt
astroid==2.1.0
beautifulsoup4==4.7.1
bs4==0.0.1
CacheControl==0.12.5
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
firebase-admin==2.16.0
Flask==1.0.2
Flask-MySQL==1.4.0
gcloud==0.17.0
google-api-core==1.8.2
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-firestore==0.32.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
grpcio==1.19.0
httplib2==0.12.1
idna==2.8
imutils==0.5.2
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jws==0.1.3
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.1
numpy==1.16.2
oauth2client==3.0.0
opencv-python==3.3.0.9
pandas==0.24.2
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycryptodome==3.4.3
pylint==2.2.2
PyMySQL==0.9.3
Pyrebase==3.0.27
python-dateutil==2.8.0
python-jwt==2.0.1
pytz==2018.9
requests==2.21.0
requests-toolbelt==0.7.0
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
typed-ast==1.3.1
urllib3==1.24.1
Werkzeug==0.15.1
wincertstore==0.2
wrapt==1.11.1
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
Pillow==5.4.1



gunicorn==19.9.0